feat: add number of concurrent requests parameter to load and verify (#25)
Adds and implements a parameter to the load and verify commands to specify the maximum number of concurrent requests.

Previously these commands processed a file of data sequentially, awaiting one request at a time. To achieve parallelism before, we ran multiple processes in the load-many script. While OK for a one time POC, that behavior is not easy for a first time user. With this new feature, we enable that parallelism with a flag on the main program.

With this we refactor the load script to use this parameter and delete the now unnecessary load-many script.

A future PR will refactor the CLI arg parsing of the scripts to make them easier to use.