feat: run the rdbcli tool directly instead of on docker
The rdbcli tool docker container is built on linux/amd64 and runs the
JVM. I have been unable to run this container on macos, even with the
appropriate platform argument.

In order to ensure users on different platforms can run the tool, I am
bundling the tool with the distributable. Per the tool's license,
Apache 2.0, this is OK as the tool comes with the license.